Added TYP_pending pseudo type + Ir instruction comments.
An improvement to the test framework: test function that can inspect the
compiler context after compilation.

TYP_pending types are to be inferred later. The type has base_type set,
if the type has been inferred. There is, however, a problem: if the type
thus far has been inferred to be TYP_null, this would cause erroneous
errors in case null type was used in an context where it normally was an
error. The final inferred type could still be pending and inferrable
after more statements, thus making the type into TYP_pointer of some
kind.