Added bintrees.hp sample, fixed a couple bugs..
Codegen bug: add/sub stack pointer was treated as a def (write) and not
modify (read-write), which resulted in subsequent instructions to seem
like two defs of the rsp to the optimizer. The optimizer would comment
out one of them.

Builtin symbols bug: builtin symbols definition file location was
NoFileLocation(). This would cause error when trying to print error
message with the file location. Now using a "special" file location with
filename "<builtin>" for builtins.